requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.3.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
exceptiongroup==1.2.2
fastapi==0.115.5
Flask==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mock==5.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
